moved the stateless functions to session utils module
